+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 [Unreleased]: https://github.com/LoicGrobol/ginger/compare/v0.7.0...HEAD
+### Fixed
+  - Remove obsolete tests
+
+### Changed
+  - Hardened CoNLL parsing strictness
+  - Moved `ParsingError` from `libginger` to `libtreebank`
 
 ## [0.7.0] - 2017-04-07
 [0.7.0]: https://github.com/LoicGrobol/ginger/compare/v0.6.0...v0.7.0